+++ src/editor/edit.h
@@ -62,11 +62,11 @@ extern int show_right_margin;
 void edit_stack_init (void);
 void edit_stack_free (void);
 
-gboolean edit_file (const vfs_path_t * file_vpath, int line);
+gboolean edit_file (const vfs_path_t * file_vpath, long line);
 gboolean edit_files (const GList * files);
 
 char *edit_get_file_name (const WEdit * edit);
-int edit_get_curs_col (const WEdit * edit);
+long edit_get_curs_col (const WEdit * edit);
 const char *edit_get_syntax_type (const WEdit * edit);
 
 /*** inline functions ****************************************************************************/
