+++ src/editor/editwidget.c
@@ -1069,7 +1069,7 @@ edit_callback (Widget * w, widget_msg_t msg, int parm)
  */
 
 gboolean
-edit_file (const vfs_path_t * file_vpath, int line)
+edit_file (const vfs_path_t * file_vpath, long line)
 {
     mcedit_arg_t arg = { (vfs_path_t *) file_vpath, line };
     GList *files;
@@ -1232,7 +1232,7 @@ edit_save_size (WEdit * edit)
  */
 
 gboolean
-edit_add_window (Dlg_head * h, int y, int x, int lines, int cols, const vfs_path_t * f, int fline)
+edit_add_window (Dlg_head * h, int y, int x, int lines, int cols, const vfs_path_t * f, long fline)
 {
     WEdit *edit;
     Widget *w;
